Integrated custom output resolution
Default behaviour was to save the stitched end result on maximum resolution. This commit provides the user with the ability to specify the resolution (µm/pixel) at which the final result should be saved. Writing speed of the final result has also greatly improved by more efficiently computing the final transformations.