added csv output of GA fitness
The genetic algorithm now outputs the scaled cost per iteration in a CSV file for easy reference. This allows the user to gauge how the genetic algorithm improved the stitch over the different resolutions. The y-axis is scaled to the absolute error before the genetic algorithm is employed. Hence, the error starts at 1 and decreases over the resolution once the error becomes a fraction of the starting situation.