added manual config 4 fragments
Added the possibility to manually specify the orientation of all fragments for a given patient in case the automatic fragment classification fails. This was previously already possible for 2 fragments, but is now possible for 4 fragments as well. Currently the four options include "ul" (upper left), "ur" (upper right), "ll" (lower left) and "lr" (lower right).